+++ test/test_ndb_main.cpp
@@ -214,7 +214,7 @@ public:
           size_t int_period(period_guess);
           double first_ac(0);
           for (ssize_t i(1); i<100 && period_guess/i > min_index; ++i) {
-            ASSERT_THROW(period_guess/i < ac.size());
+            ASSERT_THROW(size_t(period_guess/i) < ac.size());
 
             const double current_ac(ac[period_guess/i]);
             if (i == 1)
